refactor: remove modifiers and use tree inputs
