+++ nodes/collection_props.py
@@ -3,7 +3,7 @@ from bpy.types import Node
 
 from .base import FNBaseNode
 from ..sockets import FNSocketCollection, FNSocketBool
-from ..operators import get_active_mod_item
+
 
 
 class FNCollectionProps(Node, FNBaseNode):
@@ -25,9 +25,9 @@ class FNCollectionProps(Node, FNBaseNode):
         coll = inputs.get("Collection")
         if coll:
             hide_vp = inputs.get("Hide Viewport")
-            mod = get_active_mod_item()
-            if mod:
-                mod.store_original(coll, "hide_viewport")
+            ctx = getattr(getattr(self, "id_data", None), "fn_inputs", None)
+            if ctx:
+                ctx.store_original(coll, "hide_viewport")
             try:
                 coll.hide_viewport = hide_vp
             except Exception:
