+++ nodes/output_props.py
@@ -3,7 +3,7 @@ from bpy.types import Node
 
 from .base import FNBaseNode
 from ..sockets import FNSocketScene, FNSocketInt
-from ..operators import get_active_mod_item
+
 
 
 class FNOutputProps(Node, FNBaseNode):
@@ -28,10 +28,10 @@ class FNOutputProps(Node, FNBaseNode):
         if scene:
             res_x = inputs.get("Resolution X")
             res_y = inputs.get("Resolution Y")
-            mod = get_active_mod_item()
-            if mod:
-                mod.store_original(scene.render, "resolution_x")
-                mod.store_original(scene.render, "resolution_y")
+            ctx = getattr(getattr(self, "id_data", None), "fn_inputs", None)
+            if ctx:
+                ctx.store_original(scene.render, "resolution_x")
+                ctx.store_original(scene.render, "resolution_y")
             try:
                 scene.render.resolution_x = res_x
                 scene.render.resolution_y = res_y
