+++ ui.py
@@ -1,7 +1,6 @@
 
 import bpy
 from bpy.types import Panel
-from .modifiers import FILE_NODES_UL_modifiers, FileNodeModItem
 from . import ADDON_NAME
 
 class FILE_NODES_PT_global(Panel):
@@ -12,22 +11,6 @@ class FILE_NODES_PT_global(Panel):
 
     def draw(self, context):
         layout = self.layout
-        scene = context.scene
-        layout.template_list("FILE_NODES_UL_modifiers", "", scene, "file_node_modifiers", scene, "file_node_mod_index")
-        row = layout.row(align=True)
-        row.operator('file_nodes.mod_add', text="", icon='ADD')
-        row.operator('file_nodes.mod_remove', text="", icon='REMOVE')
-        row.operator('file_nodes.mod_move', text="", icon='TRIA_UP').direction = 'UP'
-        row.operator('file_nodes.mod_move', text="", icon='TRIA_DOWN').direction = 'DOWN'
-
-        if 0 <= scene.file_node_mod_index < len(scene.file_node_modifiers):
-            mod = scene.file_node_modifiers[scene.file_node_mod_index]
-            box = layout.box()
-            for inp in mod.inputs:
-                prop = inp.prop_name()
-                if prop:
-                    box.prop(inp, prop, text=inp.name)
-
         layout.operator('file_nodes.evaluate', icon='FILE_REFRESH')
         prefs = context.preferences.addons[ADDON_NAME].preferences
         layout.prop(prefs, "auto_evaluate")
